Trait initial_conditions::utils::vec_traits::ScalarDiv

pub trait ScalarDiv<T> {
    fn scalar_div(&mut self, k: T)
    where
        T: Float;
}

scalar_div (Trait)

divides each item in Vec by k, mutating self

If zero is passed as k, value will be inf

Required Methods

fn scalar_div(&mut self, k: T) where
T: Float,